requirements.txt
aiohttp==3.7.4.post0
asgiref==3.2.5
astroid==2.4.2
asttokens==2.0.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
backtrader==1.9.76.123
binance==0.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
dateparser==1.0.0
debugpy==1.6.2
decorator==5.1.1
Django==3.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.10.0
h5py==2.10.0
idna==2.10
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.15.1
ipython==8.4.0
isort==5.7.0
jedi==0.18.1
joblib==0.15.1
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.6.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.41.1
matplotlib==3.2.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==5.1.0
nest-asyncio==1.5.5
numba==0.58.1
numpy==1.24.4
oauthlib==3.1.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
plotly==5.3.1
pluggy==1.3.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
PyAlgoTrade==0.20
pyaml==21.8.3
Pygments==2.12.0
pylint==2.6.2
pyparsing==2.4.7
pytest==7.4.3
pytest-asyncio==0.23.2
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==5.3.1
pyzmq==23.2.1
reactor @ file:///C:/Users/iyede/AppData/Local/Jenkins/.jenkins/workspace/reactor_1
regex==2021.8.28
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
stack-data==0.4.0
tenacity==8.0.1
threadpoolctl==2.0.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.3.0
tweepy==3.10.0
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.3
wcwidth==0.2.5
websockets==9.1
wrapt==1.12.1
ws4py==0.5.1
yarl==1.6.3
zipp==3.17.0
